Delete LookupResources v1, ReachableResources and all helper code
SpiceDB no longer dispatches to these methods, so we can remove them
