+++ internal/dispatch/caching/delegate.go
@@ -28,14 +28,6 @@ func (fd fakeDelegate) DispatchExpand(_ context.Context, _ *v1.DispatchExpandReq
 	return &v1.DispatchExpandResponse{}, spiceerrors.MustBugf(errMessage)
 }
 
-func (fd fakeDelegate) DispatchReachableResources(_ *v1.DispatchReachableResourcesRequest, _ dispatch.ReachableResourcesStream) error {
-	return spiceerrors.MustBugf(errMessage)
-}
-
-func (fd fakeDelegate) DispatchLookupResources(_ *v1.DispatchLookupResourcesRequest, _ dispatch.LookupResourcesStream) error {
-	return spiceerrors.MustBugf(errMessage)
-}
-
 func (fd fakeDelegate) DispatchLookupResources2(_ *v1.DispatchLookupResources2Request, _ dispatch.LookupResources2Stream) error {
 	return spiceerrors.MustBugf(errMessage)
 }
