+++ internal/dispatch/keys/keys.go
@@ -13,9 +13,6 @@ type Handler interface {
 	// CheckCacheKey computes the caching key for a Check operation.
 	CheckCacheKey(ctx context.Context, req *v1.DispatchCheckRequest) (DispatchCacheKey, error)
 
-	// LookupResourcesCacheKey computes the caching key for a LookupResources operation.
-	LookupResourcesCacheKey(ctx context.Context, req *v1.DispatchLookupResourcesRequest) (DispatchCacheKey, error)
-
 	// LookupResources2CacheKey computes the caching key for a LookupResources2 operation.
 	LookupResources2CacheKey(ctx context.Context, req *v1.DispatchLookupResources2Request) (DispatchCacheKey, error)
 
@@ -25,14 +22,8 @@ type Handler interface {
 	// ExpandCacheKey computes the caching key for an Expand operation.
 	ExpandCacheKey(ctx context.Context, req *v1.DispatchExpandRequest) (DispatchCacheKey, error)
 
-	// ReachableResourcesCacheKey computes the caching key for a ReachableResources operation.
-	ReachableResourcesCacheKey(ctx context.Context, req *v1.DispatchReachableResourcesRequest) (DispatchCacheKey, error)
-
 	// CheckDispatchKey computes the dispatch key for a Check operation.
 	CheckDispatchKey(ctx context.Context, req *v1.DispatchCheckRequest) ([]byte, error)
-
-	// LookupResourcesDispatchKey computes the dispatch key for a LookupResources operation.
-	LookupResourcesDispatchKey(ctx context.Context, req *v1.DispatchLookupResourcesRequest) ([]byte, error)
 
 	// LookupResources2DispatchKey computes the dispatch key for a LookupResources2 operation.
 	LookupResources2DispatchKey(ctx context.Context, req *v1.DispatchLookupResources2Request) ([]byte, error)
@@ -42,16 +33,9 @@ type Handler interface {
 
 	// ExpandDispatchKey computes the dispatch key for an Expand operation.
 	ExpandDispatchKey(ctx context.Context, req *v1.DispatchExpandRequest) ([]byte, error)
-
-	// ReachableResourcesDispatchKey computes the key for a ReachableResources operation.
-	ReachableResourcesDispatchKey(ctx context.Context, req *v1.DispatchReachableResourcesRequest) ([]byte, error)
 }
 
 type baseKeyHandler struct{}
-
-func (b baseKeyHandler) LookupResourcesCacheKey(_ context.Context, req *v1.DispatchLookupResourcesRequest) (DispatchCacheKey, error) {
-	return lookupResourcesRequestToKey(req, computeBothHashes), nil
-}
 
 func (b baseKeyHandler) LookupResources2CacheKey(_ context.Context, req *v1.DispatchLookupResources2Request) (DispatchCacheKey, error) {
 	return lookupResourcesRequest2ToKey(req, computeBothHashes), nil
@@ -65,16 +49,8 @@ func (b baseKeyHandler) ExpandCacheKey(_ context.Context, req *v1.DispatchExpand
 	return expandRequestToKey(req, computeBothHashes), nil
 }
 
-func (b baseKeyHandler) ReachableResourcesCacheKey(_ context.Context, req *v1.DispatchReachableResourcesRequest) (DispatchCacheKey, error) {
-	return reachableResourcesRequestToKey(req, computeBothHashes), nil
-}
-
 func (b baseKeyHandler) CheckDispatchKey(_ context.Context, req *v1.DispatchCheckRequest) ([]byte, error) {
 	return checkRequestToKey(req, computeOnlyStableHash).StableSumAsBytes(), nil
-}
-
-func (b baseKeyHandler) LookupResourcesDispatchKey(_ context.Context, req *v1.DispatchLookupResourcesRequest) ([]byte, error) {
-	return lookupResourcesRequestToKey(req, computeOnlyStableHash).StableSumAsBytes(), nil
 }
 
 func (b baseKeyHandler) LookupResources2DispatchKey(_ context.Context, req *v1.DispatchLookupResources2Request) ([]byte, error) {
@@ -87,10 +63,6 @@ func (b baseKeyHandler) LookupSubjectsDispatchKey(_ context.Context, req *v1.Dis
 
 func (b baseKeyHandler) ExpandDispatchKey(_ context.Context, req *v1.DispatchExpandRequest) ([]byte, error) {
 	return expandRequestToKey(req, computeOnlyStableHash).StableSumAsBytes(), nil
-}
-
-func (b baseKeyHandler) ReachableResourcesDispatchKey(_ context.Context, req *v1.DispatchReachableResourcesRequest) ([]byte, error) {
-	return reachableResourcesRequestToKey(req, computeOnlyStableHash).StableSumAsBytes(), nil
 }
 
 // DirectKeyHandler is a key handler that uses the relation name itself as the key.
