+++ pkg/proto/dispatch/v1/dispatch.pb.validate.go
@@ -1854,1137 +1854,6 @@ var _ interface {
 	ErrorName() string
 } = DispatchLookupResources2ResponseValidationError{}
 
-// Validate checks the field values on DispatchReachableResourcesRequest with
-// the rules defined in the proto definition for this message. If any rules
-// are violated, the first error encountered is returned, or nil if there are
-// no violations.
-func (m *DispatchReachableResourcesRequest) Validate() error {
-	return m.validate(false)
-}
-
-// ValidateAll checks the field values on DispatchReachableResourcesRequest
-// with the rules defined in the proto definition for this message. If any
-// rules are violated, the result is a list of violation errors wrapped in
-// DispatchReachableResourcesRequestMultiError, or nil if none found.
-func (m *DispatchReachableResourcesRequest) ValidateAll() error {
-	return m.validate(true)
-}
-
-func (m *DispatchReachableResourcesRequest) validate(all bool) error {
-	if m == nil {
-		return nil
-	}
-
-	var errors []error
-
-	if m.GetMetadata() == nil {
-		err := DispatchReachableResourcesRequestValidationError{
-			field:  "Metadata",
-			reason: "value is required",
-		}
-		if !all {
-			return err
-		}
-		errors = append(errors, err)
-	}
-
-	if all {
-		switch v := interface{}(m.GetMetadata()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchReachableResourcesRequestValidationError{
-					field:  "Metadata",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchReachableResourcesRequestValidationError{
-					field:  "Metadata",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetMetadata()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchReachableResourcesRequestValidationError{
-				field:  "Metadata",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if m.GetResourceRelation() == nil {
-		err := DispatchReachableResourcesRequestValidationError{
-			field:  "ResourceRelation",
-			reason: "value is required",
-		}
-		if !all {
-			return err
-		}
-		errors = append(errors, err)
-	}
-
-	if all {
-		switch v := interface{}(m.GetResourceRelation()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchReachableResourcesRequestValidationError{
-					field:  "ResourceRelation",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchReachableResourcesRequestValidationError{
-					field:  "ResourceRelation",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetResourceRelation()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchReachableResourcesRequestValidationError{
-				field:  "ResourceRelation",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if m.GetSubjectRelation() == nil {
-		err := DispatchReachableResourcesRequestValidationError{
-			field:  "SubjectRelation",
-			reason: "value is required",
-		}
-		if !all {
-			return err
-		}
-		errors = append(errors, err)
-	}
-
-	if all {
-		switch v := interface{}(m.GetSubjectRelation()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchReachableResourcesRequestValidationError{
-					field:  "SubjectRelation",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchReachableResourcesRequestValidationError{
-					field:  "SubjectRelation",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetSubjectRelation()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchReachableResourcesRequestValidationError{
-				field:  "SubjectRelation",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if all {
-		switch v := interface{}(m.GetOptionalCursor()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchReachableResourcesRequestValidationError{
-					field:  "OptionalCursor",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchReachableResourcesRequestValidationError{
-					field:  "OptionalCursor",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetOptionalCursor()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchReachableResourcesRequestValidationError{
-				field:  "OptionalCursor",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	// no validation rules for OptionalLimit
-
-	if len(errors) > 0 {
-		return DispatchReachableResourcesRequestMultiError(errors)
-	}
-
-	return nil
-}
-
-// DispatchReachableResourcesRequestMultiError is an error wrapping multiple
-// validation errors returned by
-// DispatchReachableResourcesRequest.ValidateAll() if the designated
-// constraints aren't met.
-type DispatchReachableResourcesRequestMultiError []error
-
-// Error returns a concatenation of all the error messages it wraps.
-func (m DispatchReachableResourcesRequestMultiError) Error() string {
-	var msgs []string
-	for _, err := range m {
-		msgs = append(msgs, err.Error())
-	}
-	return strings.Join(msgs, "; ")
-}
-
-// AllErrors returns a list of validation violation errors.
-func (m DispatchReachableResourcesRequestMultiError) AllErrors() []error { return m }
-
-// DispatchReachableResourcesRequestValidationError is the validation error
-// returned by DispatchReachableResourcesRequest.Validate if the designated
-// constraints aren't met.
-type DispatchReachableResourcesRequestValidationError struct {
-	field  string
-	reason string
-	cause  error
-	key    bool
-}
-
-// Field function returns field value.
-func (e DispatchReachableResourcesRequestValidationError) Field() string { return e.field }
-
-// Reason function returns reason value.
-func (e DispatchReachableResourcesRequestValidationError) Reason() string { return e.reason }
-
-// Cause function returns cause value.
-func (e DispatchReachableResourcesRequestValidationError) Cause() error { return e.cause }
-
-// Key function returns key value.
-func (e DispatchReachableResourcesRequestValidationError) Key() bool { return e.key }
-
-// ErrorName returns error name.
-func (e DispatchReachableResourcesRequestValidationError) ErrorName() string {
-	return "DispatchReachableResourcesRequestValidationError"
-}
-
-// Error satisfies the builtin error interface
-func (e DispatchReachableResourcesRequestValidationError) Error() string {
-	cause := ""
-	if e.cause != nil {
-		cause = fmt.Sprintf(" | caused by: %v", e.cause)
-	}
-
-	key := ""
-	if e.key {
-		key = "key for "
-	}
-
-	return fmt.Sprintf(
-		"invalid %sDispatchReachableResourcesRequest.%s: %s%s",
-		key,
-		e.field,
-		e.reason,
-		cause)
-}
-
-var _ error = DispatchReachableResourcesRequestValidationError{}
-
-var _ interface {
-	Field() string
-	Reason() string
-	Key() bool
-	Cause() error
-	ErrorName() string
-} = DispatchReachableResourcesRequestValidationError{}
-
-// Validate checks the field values on ReachableResource with the rules defined
-// in the proto definition for this message. If any rules are violated, the
-// first error encountered is returned, or nil if there are no violations.
-func (m *ReachableResource) Validate() error {
-	return m.validate(false)
-}
-
-// ValidateAll checks the field values on ReachableResource with the rules
-// defined in the proto definition for this message. If any rules are
-// violated, the result is a list of violation errors wrapped in
-// ReachableResourceMultiError, or nil if none found.
-func (m *ReachableResource) ValidateAll() error {
-	return m.validate(true)
-}
-
-func (m *ReachableResource) validate(all bool) error {
-	if m == nil {
-		return nil
-	}
-
-	var errors []error
-
-	// no validation rules for ResourceId
-
-	// no validation rules for ResultStatus
-
-	if len(errors) > 0 {
-		return ReachableResourceMultiError(errors)
-	}
-
-	return nil
-}
-
-// ReachableResourceMultiError is an error wrapping multiple validation errors
-// returned by ReachableResource.ValidateAll() if the designated constraints
-// aren't met.
-type ReachableResourceMultiError []error
-
-// Error returns a concatenation of all the error messages it wraps.
-func (m ReachableResourceMultiError) Error() string {
-	var msgs []string
-	for _, err := range m {
-		msgs = append(msgs, err.Error())
-	}
-	return strings.Join(msgs, "; ")
-}
-
-// AllErrors returns a list of validation violation errors.
-func (m ReachableResourceMultiError) AllErrors() []error { return m }
-
-// ReachableResourceValidationError is the validation error returned by
-// ReachableResource.Validate if the designated constraints aren't met.
-type ReachableResourceValidationError struct {
-	field  string
-	reason string
-	cause  error
-	key    bool
-}
-
-// Field function returns field value.
-func (e ReachableResourceValidationError) Field() string { return e.field }
-
-// Reason function returns reason value.
-func (e ReachableResourceValidationError) Reason() string { return e.reason }
-
-// Cause function returns cause value.
-func (e ReachableResourceValidationError) Cause() error { return e.cause }
-
-// Key function returns key value.
-func (e ReachableResourceValidationError) Key() bool { return e.key }
-
-// ErrorName returns error name.
-func (e ReachableResourceValidationError) ErrorName() string {
-	return "ReachableResourceValidationError"
-}
-
-// Error satisfies the builtin error interface
-func (e ReachableResourceValidationError) Error() string {
-	cause := ""
-	if e.cause != nil {
-		cause = fmt.Sprintf(" | caused by: %v", e.cause)
-	}
-
-	key := ""
-	if e.key {
-		key = "key for "
-	}
-
-	return fmt.Sprintf(
-		"invalid %sReachableResource.%s: %s%s",
-		key,
-		e.field,
-		e.reason,
-		cause)
-}
-
-var _ error = ReachableResourceValidationError{}
-
-var _ interface {
-	Field() string
-	Reason() string
-	Key() bool
-	Cause() error
-	ErrorName() string
-} = ReachableResourceValidationError{}
-
-// Validate checks the field values on DispatchReachableResourcesResponse with
-// the rules defined in the proto definition for this message. If any rules
-// are violated, the first error encountered is returned, or nil if there are
-// no violations.
-func (m *DispatchReachableResourcesResponse) Validate() error {
-	return m.validate(false)
-}
-
-// ValidateAll checks the field values on DispatchReachableResourcesResponse
-// with the rules defined in the proto definition for this message. If any
-// rules are violated, the result is a list of violation errors wrapped in
-// DispatchReachableResourcesResponseMultiError, or nil if none found.
-func (m *DispatchReachableResourcesResponse) ValidateAll() error {
-	return m.validate(true)
-}
-
-func (m *DispatchReachableResourcesResponse) validate(all bool) error {
-	if m == nil {
-		return nil
-	}
-
-	var errors []error
-
-	if all {
-		switch v := interface{}(m.GetResource()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchReachableResourcesResponseValidationError{
-					field:  "Resource",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchReachableResourcesResponseValidationError{
-					field:  "Resource",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetResource()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchReachableResourcesResponseValidationError{
-				field:  "Resource",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if all {
-		switch v := interface{}(m.GetMetadata()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchReachableResourcesResponseValidationError{
-					field:  "Metadata",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchReachableResourcesResponseValidationError{
-					field:  "Metadata",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetMetadata()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchReachableResourcesResponseValidationError{
-				field:  "Metadata",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if all {
-		switch v := interface{}(m.GetAfterResponseCursor()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchReachableResourcesResponseValidationError{
-					field:  "AfterResponseCursor",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchReachableResourcesResponseValidationError{
-					field:  "AfterResponseCursor",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetAfterResponseCursor()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchReachableResourcesResponseValidationError{
-				field:  "AfterResponseCursor",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if len(errors) > 0 {
-		return DispatchReachableResourcesResponseMultiError(errors)
-	}
-
-	return nil
-}
-
-// DispatchReachableResourcesResponseMultiError is an error wrapping multiple
-// validation errors returned by
-// DispatchReachableResourcesResponse.ValidateAll() if the designated
-// constraints aren't met.
-type DispatchReachableResourcesResponseMultiError []error
-
-// Error returns a concatenation of all the error messages it wraps.
-func (m DispatchReachableResourcesResponseMultiError) Error() string {
-	var msgs []string
-	for _, err := range m {
-		msgs = append(msgs, err.Error())
-	}
-	return strings.Join(msgs, "; ")
-}
-
-// AllErrors returns a list of validation violation errors.
-func (m DispatchReachableResourcesResponseMultiError) AllErrors() []error { return m }
-
-// DispatchReachableResourcesResponseValidationError is the validation error
-// returned by DispatchReachableResourcesResponse.Validate if the designated
-// constraints aren't met.
-type DispatchReachableResourcesResponseValidationError struct {
-	field  string
-	reason string
-	cause  error
-	key    bool
-}
-
-// Field function returns field value.
-func (e DispatchReachableResourcesResponseValidationError) Field() string { return e.field }
-
-// Reason function returns reason value.
-func (e DispatchReachableResourcesResponseValidationError) Reason() string { return e.reason }
-
-// Cause function returns cause value.
-func (e DispatchReachableResourcesResponseValidationError) Cause() error { return e.cause }
-
-// Key function returns key value.
-func (e DispatchReachableResourcesResponseValidationError) Key() bool { return e.key }
-
-// ErrorName returns error name.
-func (e DispatchReachableResourcesResponseValidationError) ErrorName() string {
-	return "DispatchReachableResourcesResponseValidationError"
-}
-
-// Error satisfies the builtin error interface
-func (e DispatchReachableResourcesResponseValidationError) Error() string {
-	cause := ""
-	if e.cause != nil {
-		cause = fmt.Sprintf(" | caused by: %v", e.cause)
-	}
-
-	key := ""
-	if e.key {
-		key = "key for "
-	}
-
-	return fmt.Sprintf(
-		"invalid %sDispatchReachableResourcesResponse.%s: %s%s",
-		key,
-		e.field,
-		e.reason,
-		cause)
-}
-
-var _ error = DispatchReachableResourcesResponseValidationError{}
-
-var _ interface {
-	Field() string
-	Reason() string
-	Key() bool
-	Cause() error
-	ErrorName() string
-} = DispatchReachableResourcesResponseValidationError{}
-
-// Validate checks the field values on DispatchLookupResourcesRequest with the
-// rules defined in the proto definition for this message. If any rules are
-// violated, the first error encountered is returned, or nil if there are no violations.
-func (m *DispatchLookupResourcesRequest) Validate() error {
-	return m.validate(false)
-}
-
-// ValidateAll checks the field values on DispatchLookupResourcesRequest with
-// the rules defined in the proto definition for this message. If any rules
-// are violated, the result is a list of violation errors wrapped in
-// DispatchLookupResourcesRequestMultiError, or nil if none found.
-func (m *DispatchLookupResourcesRequest) ValidateAll() error {
-	return m.validate(true)
-}
-
-func (m *DispatchLookupResourcesRequest) validate(all bool) error {
-	if m == nil {
-		return nil
-	}
-
-	var errors []error
-
-	if m.GetMetadata() == nil {
-		err := DispatchLookupResourcesRequestValidationError{
-			field:  "Metadata",
-			reason: "value is required",
-		}
-		if !all {
-			return err
-		}
-		errors = append(errors, err)
-	}
-
-	if all {
-		switch v := interface{}(m.GetMetadata()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "Metadata",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "Metadata",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetMetadata()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchLookupResourcesRequestValidationError{
-				field:  "Metadata",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if m.GetObjectRelation() == nil {
-		err := DispatchLookupResourcesRequestValidationError{
-			field:  "ObjectRelation",
-			reason: "value is required",
-		}
-		if !all {
-			return err
-		}
-		errors = append(errors, err)
-	}
-
-	if all {
-		switch v := interface{}(m.GetObjectRelation()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "ObjectRelation",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "ObjectRelation",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetObjectRelation()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchLookupResourcesRequestValidationError{
-				field:  "ObjectRelation",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if m.GetSubject() == nil {
-		err := DispatchLookupResourcesRequestValidationError{
-			field:  "Subject",
-			reason: "value is required",
-		}
-		if !all {
-			return err
-		}
-		errors = append(errors, err)
-	}
-
-	if all {
-		switch v := interface{}(m.GetSubject()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "Subject",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "Subject",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetSubject()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchLookupResourcesRequestValidationError{
-				field:  "Subject",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if all {
-		switch v := interface{}(m.GetContext()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "Context",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "Context",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetContext()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchLookupResourcesRequestValidationError{
-				field:  "Context",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	// no validation rules for OptionalLimit
-
-	if all {
-		switch v := interface{}(m.GetOptionalCursor()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "OptionalCursor",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchLookupResourcesRequestValidationError{
-					field:  "OptionalCursor",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetOptionalCursor()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchLookupResourcesRequestValidationError{
-				field:  "OptionalCursor",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if len(errors) > 0 {
-		return DispatchLookupResourcesRequestMultiError(errors)
-	}
-
-	return nil
-}
-
-// DispatchLookupResourcesRequestMultiError is an error wrapping multiple
-// validation errors returned by DispatchLookupResourcesRequest.ValidateAll()
-// if the designated constraints aren't met.
-type DispatchLookupResourcesRequestMultiError []error
-
-// Error returns a concatenation of all the error messages it wraps.
-func (m DispatchLookupResourcesRequestMultiError) Error() string {
-	var msgs []string
-	for _, err := range m {
-		msgs = append(msgs, err.Error())
-	}
-	return strings.Join(msgs, "; ")
-}
-
-// AllErrors returns a list of validation violation errors.
-func (m DispatchLookupResourcesRequestMultiError) AllErrors() []error { return m }
-
-// DispatchLookupResourcesRequestValidationError is the validation error
-// returned by DispatchLookupResourcesRequest.Validate if the designated
-// constraints aren't met.
-type DispatchLookupResourcesRequestValidationError struct {
-	field  string
-	reason string
-	cause  error
-	key    bool
-}
-
-// Field function returns field value.
-func (e DispatchLookupResourcesRequestValidationError) Field() string { return e.field }
-
-// Reason function returns reason value.
-func (e DispatchLookupResourcesRequestValidationError) Reason() string { return e.reason }
-
-// Cause function returns cause value.
-func (e DispatchLookupResourcesRequestValidationError) Cause() error { return e.cause }
-
-// Key function returns key value.
-func (e DispatchLookupResourcesRequestValidationError) Key() bool { return e.key }
-
-// ErrorName returns error name.
-func (e DispatchLookupResourcesRequestValidationError) ErrorName() string {
-	return "DispatchLookupResourcesRequestValidationError"
-}
-
-// Error satisfies the builtin error interface
-func (e DispatchLookupResourcesRequestValidationError) Error() string {
-	cause := ""
-	if e.cause != nil {
-		cause = fmt.Sprintf(" | caused by: %v", e.cause)
-	}
-
-	key := ""
-	if e.key {
-		key = "key for "
-	}
-
-	return fmt.Sprintf(
-		"invalid %sDispatchLookupResourcesRequest.%s: %s%s",
-		key,
-		e.field,
-		e.reason,
-		cause)
-}
-
-var _ error = DispatchLookupResourcesRequestValidationError{}
-
-var _ interface {
-	Field() string
-	Reason() string
-	Key() bool
-	Cause() error
-	ErrorName() string
-} = DispatchLookupResourcesRequestValidationError{}
-
-// Validate checks the field values on ResolvedResource with the rules defined
-// in the proto definition for this message. If any rules are violated, the
-// first error encountered is returned, or nil if there are no violations.
-func (m *ResolvedResource) Validate() error {
-	return m.validate(false)
-}
-
-// ValidateAll checks the field values on ResolvedResource with the rules
-// defined in the proto definition for this message. If any rules are
-// violated, the result is a list of violation errors wrapped in
-// ResolvedResourceMultiError, or nil if none found.
-func (m *ResolvedResource) ValidateAll() error {
-	return m.validate(true)
-}
-
-func (m *ResolvedResource) validate(all bool) error {
-	if m == nil {
-		return nil
-	}
-
-	var errors []error
-
-	// no validation rules for ResourceId
-
-	// no validation rules for Permissionship
-
-	if len(errors) > 0 {
-		return ResolvedResourceMultiError(errors)
-	}
-
-	return nil
-}
-
-// ResolvedResourceMultiError is an error wrapping multiple validation errors
-// returned by ResolvedResource.ValidateAll() if the designated constraints
-// aren't met.
-type ResolvedResourceMultiError []error
-
-// Error returns a concatenation of all the error messages it wraps.
-func (m ResolvedResourceMultiError) Error() string {
-	var msgs []string
-	for _, err := range m {
-		msgs = append(msgs, err.Error())
-	}
-	return strings.Join(msgs, "; ")
-}
-
-// AllErrors returns a list of validation violation errors.
-func (m ResolvedResourceMultiError) AllErrors() []error { return m }
-
-// ResolvedResourceValidationError is the validation error returned by
-// ResolvedResource.Validate if the designated constraints aren't met.
-type ResolvedResourceValidationError struct {
-	field  string
-	reason string
-	cause  error
-	key    bool
-}
-
-// Field function returns field value.
-func (e ResolvedResourceValidationError) Field() string { return e.field }
-
-// Reason function returns reason value.
-func (e ResolvedResourceValidationError) Reason() string { return e.reason }
-
-// Cause function returns cause value.
-func (e ResolvedResourceValidationError) Cause() error { return e.cause }
-
-// Key function returns key value.
-func (e ResolvedResourceValidationError) Key() bool { return e.key }
-
-// ErrorName returns error name.
-func (e ResolvedResourceValidationError) ErrorName() string { return "ResolvedResourceValidationError" }
-
-// Error satisfies the builtin error interface
-func (e ResolvedResourceValidationError) Error() string {
-	cause := ""
-	if e.cause != nil {
-		cause = fmt.Sprintf(" | caused by: %v", e.cause)
-	}
-
-	key := ""
-	if e.key {
-		key = "key for "
-	}
-
-	return fmt.Sprintf(
-		"invalid %sResolvedResource.%s: %s%s",
-		key,
-		e.field,
-		e.reason,
-		cause)
-}
-
-var _ error = ResolvedResourceValidationError{}
-
-var _ interface {
-	Field() string
-	Reason() string
-	Key() bool
-	Cause() error
-	ErrorName() string
-} = ResolvedResourceValidationError{}
-
-// Validate checks the field values on DispatchLookupResourcesResponse with the
-// rules defined in the proto definition for this message. If any rules are
-// violated, the first error encountered is returned, or nil if there are no violations.
-func (m *DispatchLookupResourcesResponse) Validate() error {
-	return m.validate(false)
-}
-
-// ValidateAll checks the field values on DispatchLookupResourcesResponse with
-// the rules defined in the proto definition for this message. If any rules
-// are violated, the result is a list of violation errors wrapped in
-// DispatchLookupResourcesResponseMultiError, or nil if none found.
-func (m *DispatchLookupResourcesResponse) ValidateAll() error {
-	return m.validate(true)
-}
-
-func (m *DispatchLookupResourcesResponse) validate(all bool) error {
-	if m == nil {
-		return nil
-	}
-
-	var errors []error
-
-	if all {
-		switch v := interface{}(m.GetMetadata()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchLookupResourcesResponseValidationError{
-					field:  "Metadata",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchLookupResourcesResponseValidationError{
-					field:  "Metadata",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetMetadata()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchLookupResourcesResponseValidationError{
-				field:  "Metadata",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if all {
-		switch v := interface{}(m.GetResolvedResource()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchLookupResourcesResponseValidationError{
-					field:  "ResolvedResource",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchLookupResourcesResponseValidationError{
-					field:  "ResolvedResource",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetResolvedResource()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchLookupResourcesResponseValidationError{
-				field:  "ResolvedResource",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if all {
-		switch v := interface{}(m.GetAfterResponseCursor()).(type) {
-		case interface{ ValidateAll() error }:
-			if err := v.ValidateAll(); err != nil {
-				errors = append(errors, DispatchLookupResourcesResponseValidationError{
-					field:  "AfterResponseCursor",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		case interface{ Validate() error }:
-			if err := v.Validate(); err != nil {
-				errors = append(errors, DispatchLookupResourcesResponseValidationError{
-					field:  "AfterResponseCursor",
-					reason: "embedded message failed validation",
-					cause:  err,
-				})
-			}
-		}
-	} else if v, ok := interface{}(m.GetAfterResponseCursor()).(interface{ Validate() error }); ok {
-		if err := v.Validate(); err != nil {
-			return DispatchLookupResourcesResponseValidationError{
-				field:  "AfterResponseCursor",
-				reason: "embedded message failed validation",
-				cause:  err,
-			}
-		}
-	}
-
-	if len(errors) > 0 {
-		return DispatchLookupResourcesResponseMultiError(errors)
-	}
-
-	return nil
-}
-
-// DispatchLookupResourcesResponseMultiError is an error wrapping multiple
-// validation errors returned by DispatchLookupResourcesResponse.ValidateAll()
-// if the designated constraints aren't met.
-type DispatchLookupResourcesResponseMultiError []error
-
-// Error returns a concatenation of all the error messages it wraps.
-func (m DispatchLookupResourcesResponseMultiError) Error() string {
-	var msgs []string
-	for _, err := range m {
-		msgs = append(msgs, err.Error())
-	}
-	return strings.Join(msgs, "; ")
-}
-
-// AllErrors returns a list of validation violation errors.
-func (m DispatchLookupResourcesResponseMultiError) AllErrors() []error { return m }
-
-// DispatchLookupResourcesResponseValidationError is the validation error
-// returned by DispatchLookupResourcesResponse.Validate if the designated
-// constraints aren't met.
-type DispatchLookupResourcesResponseValidationError struct {
-	field  string
-	reason string
-	cause  error
-	key    bool
-}
-
-// Field function returns field value.
-func (e DispatchLookupResourcesResponseValidationError) Field() string { return e.field }
-
-// Reason function returns reason value.
-func (e DispatchLookupResourcesResponseValidationError) Reason() string { return e.reason }
-
-// Cause function returns cause value.
-func (e DispatchLookupResourcesResponseValidationError) Cause() error { return e.cause }
-
-// Key function returns key value.
-func (e DispatchLookupResourcesResponseValidationError) Key() bool { return e.key }
-
-// ErrorName returns error name.
-func (e DispatchLookupResourcesResponseValidationError) ErrorName() string {
-	return "DispatchLookupResourcesResponseValidationError"
-}
-
-// Error satisfies the builtin error interface
-func (e DispatchLookupResourcesResponseValidationError) Error() string {
-	cause := ""
-	if e.cause != nil {
-		cause = fmt.Sprintf(" | caused by: %v", e.cause)
-	}
-
-	key := ""
-	if e.key {
-		key = "key for "
-	}
-
-	return fmt.Sprintf(
-		"invalid %sDispatchLookupResourcesResponse.%s: %s%s",
-		key,
-		e.field,
-		e.reason,
-		cause)
-}
-
-var _ error = DispatchLookupResourcesResponseValidationError{}
-
-var _ interface {
-	Field() string
-	Reason() string
-	Key() bool
-	Cause() error
-	ErrorName() string
-} = DispatchLookupResourcesResponseValidationError{}
-
 // Validate checks the field values on DispatchLookupSubjectsRequest with the
 // rules defined in the proto definition for this message. If any rules are
 // violated, the first error encountered is returned, or nil if there are no violations.
